Fixed phpstan errors

+++ app/Models/Product.php
@@ -5,6 +5,9 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
+/**
+ * @property CartProduct $pivot
+ */
 class Product extends Model
 {
     use HasFactory;
@@ -24,9 +27,9 @@ class Product extends Model
     protected static function booted(): void
     {
         static::created(function (Product $product) {
-            $stripe = new \Stripe\StripeClient(config('cashier.secret'));
+            $stripe = new \Stripe\StripeClient(['api_key' => config('cashier.secret')]);
 
-            $stripeProduct = $stripe->products->create([
+            $price = $stripe->products->create([
                 'name' => $product->name,
                 'default_price_data' => [
                     'currency' => 'eur',
@@ -34,15 +37,23 @@ class Product extends Model
                 ],
                 'images' => $product->images,
                 'expand' => ['default_price'],
-            ]);
+            ])->default_price;
+
+            if (gettype($price) != 'object' || get_class($price) !== \Stripe\Price::class) {
+                return;
+            }
 
             $product->update([
-                'stripe_id' => $stripeProduct->default_price->id,
+                'stripe_id' => $price->id,
             ]);
         });
 
         static::deleted(function (Product $product) {
-            $stripe = new \Stripe\StripeClient(config('cashier.secret'));
+            $stripe = new \Stripe\StripeClient(['api_key' => config('cashier.secret')]);
+
+            if ($product->stripe_id == null) {
+                return;
+            }
 
             $stripe->products->delete($product->stripe_id);
         });
